refactor: rename basic to basics
