+++ stories/components/Button/basics/Color/OutlinedButtonColors.tsx
