+++ stories/components/Button/basics/Color/index.tsx
